# server/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/system/notice": {
      "post": {
        "operationId": "NoticeController_create",
        "summary": "通知公告-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoticeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "通知公告"
        ]
      },
      "put": {
        "operationId": "NoticeController_update",
        "summary": "通知公告-更新",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoticeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通知公告"
        ]
      }
    },
    "/system/notice/list": {
      "get": {
        "operationId": "NoticeController_findAll",
        "summary": "通知公告-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListNoticeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通知公告"
        ]
      }
    },
    "/system/notice/{id}": {
      "get": {
        "operationId": "NoticeController_findOne",
        "summary": "通知公告-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通知公告"
        ]
      },
      "delete": {
        "operationId": "NoticeController_remove",
        "summary": "通知公告-删除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通知公告"
        ]
      }
    },
...
    "/system/tracing": {
      "post": {
        "operationId": "TracingController_create",
        "summary": "监控-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTracingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "监控"
        ]
      },
      "put": {
        "operationId": "TracingController_update",
        "summary": "监控-更新",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTracingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "监控"
        ]
      }
    },
    "/system/tracing/list": {
      "get": {
        "operationId": "TracingController_findAll",
        "summary": "监控-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListTracingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "监控"
        ]
      }
    },
    "/system/tracing/{id}": {
      "get": {
        "operationId": "TracingController_findOne",
        "summary": "监控-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "监控"
        ]
      },
      "delete": {
        "operationId": "TracingController_remove",
        "summary": "监控-删除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "监控"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DateParamsDTO": {
        "type": "object",
        "properties": {}
      },
...
      "CreateNoticeDto": {
        "type": "object",
        "properties": {
          "remark": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "remark"
        ]
      },
      "ListNoticeDto": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "number",
            "description": "当前分页",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页数量",
            "default": 10
          },
          "params": {
            "description": "时间范围",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateParamsDTO"
              }
            ]
          },
          "orderByColumn": {
            "type": "string",
            "description": "排序字段"
          },
          "isAsc": {
            "type": "string",
            "description": "排序规则"
          }
        },
        "required": [
          "pageNum",
          "pageSize"
        ]
      },
      "UpdateNoticeDto": {
        "type": "object",
        "properties": {
          "remark": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "remark"
        ]
      },
...
      "CreateTracingDto": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string"
          }
        },
        "required": [
          "eventType"
        ]
      },
      "ListTracingDto": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "number",
            "description": "当前分页",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页数量",
            "default": 10
          },
          "params": {
            "description": "时间范围",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateParamsDTO"
              }
            ]
          },
          "orderByColumn": {
            "type": "string",
            "description": "排序字段"
          },
          "isAsc": {
            "type": "string",
            "description": "排序规则"
          }
        },
        "required": [
          "pageNum",
          "pageSize"
        ]
      },
      "UpdateTracingDto": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string"
          },
          "tracingId": {
            "type": "number"
          }
        },
        "required": [
          "eventType",
          "tracingId"
        ]
      },
...
    }
  }
}